+++ R/ebv_properties.R
@@ -2,12 +2,12 @@
 #'
 #' @slot general Named list. Elements: title, description, doi, ebv_class,
 #'   ebv_name, ebv_domain, references, source, project_name, project_url,
-#'   creator_name, creator_institution, creator_email, contributor_name,
-#'   publisher_name, publisher_institution, publisher_email, comment, keywords,
-#'   id, history, licence, conventions, naming_authority, date_created,
-#'   date_issued, entity_names, entity_type, entity_scope,
-#'   entity_classification_name, entity_classification_url, taxonomy,
-#'   taxonomy_lsid, date_modified, date_metadata_modified
+#'   creator_name, creator_institution, creator_email, creator_url,
+#'   contributor_name, publisher_name, publisher_institution, publisher_email,
+#'   publisher_url, comment, keywords, id, history, licence, conventions,
+#'   naming_authority, date_created, date_issued, entity_names, entity_type,
+#'   entity_scope, entity_classification_name, entity_classification_url,
+#'   taxonomy, taxonomy_lsid, date_modified, date_metadata_modified
 #' @slot spatial Named list. Elements: wkt2, epsg, extent, resolution,
 #'   crs_units, dimensions, scope, description
 #' @slot temporal Named list. Elements: resolution, units, timesteps, dates,
@@ -210,6 +210,7 @@ ebv_properties <-
     creator_institution <-
       ebv_i_read_att(hdf, 'creator_institution', verbose)
     creator_email <- ebv_i_read_att(hdf, 'creator_email', verbose)
+    creator_url <- ebv_i_read_att(hdf, 'creator_url', verbose)
     contributor_name <-
       ebv_i_read_att(hdf, 'contributor_name', verbose)
     publisher_name <- ebv_i_read_att(hdf, 'publisher_name', verbose)
@@ -217,6 +218,7 @@ ebv_properties <-
       ebv_i_read_att(hdf, 'publisher_institution', verbose)
     publisher_email <-
       ebv_i_read_att(hdf, 'publisher_email', verbose)
+    publisher_url <- ebv_i_read_att(hdf, 'publisher_url', verbose)
     comment <- ebv_i_read_att(hdf, 'comment', verbose)
     ebv_class <- ebv_i_read_att(hdf, 'ebv_class', verbose)
     ebv_name <- ebv_i_read_att(hdf, 'ebv_name', verbose)
@@ -270,10 +272,10 @@ ebv_properties <-
     #resolution <- c(resolution, ebv_i_read_att(hdf, 'geospatial_lat_resolution'))
 
     #global spatial atts
-    ebv_spatial_scope <-
-      ebv_i_read_att(hdf, 'ebv_spatial_scope', verbose)
-    ebv_spatial_description <-
-      ebv_i_read_att(hdf, 'ebv_spatial_description', verbose)
+    ebv_geospatial_scope <-
+      ebv_i_read_att(hdf, 'ebv_geospatial_scope', verbose)
+    ebv_geospatial_description <-
+      ebv_i_read_att(hdf, 'ebv_geospatial_description', verbose)
 
     #get dims
     if (ebv_i_4D(filepath)) {
@@ -331,10 +333,12 @@ ebv_properties <-
         'creator_name' = creator_name,
         'creator_institution' = creator_institution,
         'creator_email' = creator_email,
+        'creator_url' = creator_url,
         'contributor_name' = contributor_name,
         'publisher_name' = publisher_name,
         'publisher_institution' = publisher_institution,
         'publisher_email' = publisher_email,
+        'publisher_url' = publisher_url,
         'comment' = comment,
         'keywords' = keywords,
         'id' = id,
@@ -362,8 +366,8 @@ ebv_properties <-
         'resolution' = resolution,
         'crs_units' = crs_units,
         'dimensions' = dims,
-        'scope' = ebv_spatial_scope,
-        'description' = ebv_spatial_description
+        'scope' = ebv_geospatial_scope,
+        'description' = ebv_geospatial_description
       )
     temporal <- list(
       'resolution' = t_res,